requirements.txt
python-dotenv
openai
chromadb
aiogram